requirements.txt
aiogram
python-dotenv
pyyaml
telethon
google-generativeai
